updated App.js
Now instead of using the current value to update the state we will use the call back function to update the state with the current value, and in future if there is requirement to change the value of step with some other value not the current value, then we can you the setState() function with new value without the callback requirement.

setState(s => s+1) // when you update with current value
setState(100) // when you update with some arbitary value.