Update App.js
Now think a change came down, and we are required to add emoji to the buttons showing direction.

In this we added a emoji prop and showed in the step component.

All is good, but we have requirment issue, which is saying that you have placed 'left emoji' correctly in the direction point before the text 'previous', but you haven't placed right finger 'emoji' on the right of text 'next'.

So now what can we do, shall we include one more prop called direction, to handle the direction of emoji?

And now can you call this Button component reusable? passing so many parameters, it is same thing like in java a reusable function. is that have less parameters and can called be easily.

So we need a component that requires less props and can be worked easily, so see next commit how to overcome this.

I know message is too long;/ )