Update App.js
My own implementation for Tip-Calculator, so it works as per the video pause demo showcased.

I would say, only limitation this application has when I have to add another friend, I require another useState(), so we can say that if in future if the users are added like aunt,father etc, we need to have individual state for that.

But nevertheless, I am able to build the application. :100: