Update App.js
Key-take away: We are lifting the state for friendList for the sibling components and then granting the read and write access to both, and soon there is update in friendList, both the components re-renders themselves.

onSubmit automatically passes the event object, so there is no need to explicitly mention the object during the call.

handleAddFriend() , always remember to use the previous state as parameter for the function if the new state is dependent on the previous state.