More usage of useReducer
Keytake Away: useReducer

When using the useReducer hook in React, the initial state and the state returned by the reducer should maintain the same shape to ensure consistent state management. However, if you need to update or replace the entire state object, it is possible. You can change the structure or properties of the state object within the reducer function.

-> Consistent State Shape: It's generally a good practice to keep the state shape consistent. If parts of your component expect specific properties, changing the state shape might cause runtime errors.