Answer Question & updating points
Keytake away:

1. We have answered a question with the help of new state called answer.

so when user click on the button on the option component, we dispatch the function and update the value of answer with the current button index.

And on the basis of the that we manage our css style, so in the first css conditional styling
we check if the user have answered so if index === answer, we are sure that sure clicked on that button.
And in the second style we first check is there some answered and if that is true, is the correctAnswer and index match with correctOption, although you could have used answer to check with correct option, but here index === answer, as we dispatch on that

2. Creating points, so we need another state, to manage the points, so we need to see if the user answer option and current question correct option save same value if that is the case then give the point otherwise don't update the points.