disabling the button
disabling the button if user clicked on the button.

Note to previous commit:

And we can't use answer instead of index in case of styling correct & wrong because, the

if let's say user clicked on the wrong option and we did answer === correctOption, then every button will be showing the the same color and we want the correct option should be painted with different color, so if index === correctOption is there, then we are sure that if wrong option is selected then index will match with correctOption will painted correctly and rest will be painted as wrong option.

And if the user answer correct with the help of index the correctOption will painted correctly and wrong options similarly painted differently.