Simple Reducer usage
Keytake away:

-> case "closeAccount":
      if (state.balance === 0 && state.loan === 0) return { ...initialState };

diference b/w return { ...initialState } && return { nitialState }, is that ,

when we use spread operator, This creates a new object with the same properties and values as initialState. It effectively copies the properties of initialState into a new object. As a result, when the state is updated, it triggers a re-render, and the destructuring assignment [ { balance, loan, isActive }, dispatch ] will have the new values from the initial state.

But without spread operator -

case "closeAccount":
  return { initialState };

This returns an object that has a single property initialState, which itself is an object. Instead of replacing the state with the properties of initialState, it replaces the state with an object that contains initialState as a property. This does not match the structure expected by the destructuring assignment, causing the component not to display balance and loan correctly.