Creating new city , datepicker
Keytake aways:

--> e.preventDefault() is a method in JavaScript that prevents the default behavior of the event that is being handled. For example, when applied to a form submission event, it prevents the form from being submitted and thus stops the page from refreshing or navigating away.

In the context of form submission, the default action is to send the form data to the server and reload the page. By using e.preventDefault(), you can stop this action, allowing you to handle the form data with JavaScript (e.g., via an AJAX request or updating the state in a React application).

1. Adding a city to our cities.json file with the help of newCity object we made from the form component.

2. In ContextApi, we created a createCity function that add the city to our cities file and id has not been created, its json server that creates id automatically

3. Datepicker from the react