useMemo Hook
Keytake aways:

The useMemo hook in React is used to optimize performance by memoizing the result of a computation. This means that the computation is only re-executed when its dependencies change, preventing unnecessary recalculations on every render.

Performance Optimization: Reduces the number of computations, enhancing performance, especially for heavy operations.

Avoids Unnecessary Re-renders: Helps prevent re-rendering components that depend on the result of a computation.

So in our case, we have loaded the 30K posts and then we search for the post from the search bar, we re-render all the child components including that 30K posts component(Archive component) which causes a significant delay check in Profiler tool.

So here our Archive component is heavy in terms of data, so to prevent re-render on every prop change from the parent which is not even meant for Archive there we can use memo which only renders the component when the prop required by the component changes.