useCallback hook
Keytakeaways:

1. useCallback hook is required to when memoizing a function, because like the objects the function are also created newly on each render.

2. So when a function is passed a prop to another component which will the component to render if the other state in component change, so in order to prevent that we use useCallback()

3. Note: useState setters are automatic memoized so if they are not required to passed in useEffect, useCallback, or in useMemo and eslint is smart enough to know about that.

Why to memoize the component sending memoized props should be enough?

Ans: Even if the props themselves are memoized using useMemo and useCallback, if the parent component re-renders due to changes in other state variables (like isFakeDark), the child component will still be re-rendered because React doesn't inherently know that the props haven't changed unless you use React.memo.

React.memo is a higher-order component that memoizes the result of a component's render. It will prevent the component from re-rendering if the props haven't changed.