Loading with useNavigation from react router dom
React Router with Loaders:

Loader Functions: In React Router, loader functions are used to fetch data before the route renders its component. The loader is executed as part of the navigation process, and the route waits for the data to be fetched before rendering the component.

useNavigation: The useNavigation hook tracks the state of the navigation process. If a route transition involves a loader fetching data, the navigation state will reflect that.

Process flow:

When you navigate to the /menu path, React Router will:

->Trigger the menuLoader function to fetch the necessary data.
-> While menuLoader is fetching data, the navigation state will be loading.
-> Once the data is fetched, the Menu component will render, and the navigation state will change to idle.

Keytake-aways:

1. useNavigation hook to get the the information regarding the route.
2. A addition to missed hooked in last commit useLoaderData() to get the data from the loader property for the rendered component.